requirements_web.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.3.0
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
